# Add support for deep copy