--- a/docs/examples/copy_dict.py
+++ b/docs/examples/copy_dict.py
@@ -31,3 +31,11 @@ print(m.copy(exclude={'foo', 'bar'}))
 
 print(m.copy(update={'banana': 0}))
 # > FooBarModel banana=0 foo='hello' bar=<BarModel whatever=123>
+
+print(id(m.bar), id(m.copy().bar))
+# normal copy gives the same object reference for `bar`
+# > 140494497582280 140494497582280
+
+print(id(m.bar), id(m.copy(deep=True).bar))
+# deep copy gives a new object reference for `bar`
+# > 140494497582280 140494497582856