## Change Summary

Using `model.copy(deep=True)` will deep copy a module instance.

Making a deep copy is useful for models that have lists of dicts or
any other non-simple data structure to not accidentally modify data
between instances.

`model.copy()` behaviour is unchanged and will do a shallow copy of
the model instance.

## Related issue number

Solves #249 

<!-- Are there any issues opened that will be resolved by merging this change? -->

## Performance Changes

* before: pydantic best=39.896μs/iter avg=41.037μs/iter stdev=0.729μs/iter
* after: pydantic best=40.237μs/iter avg=41.397μs/iter stdev=0.841μs/iter

## Checklist

* [x] Unit tests for the changes exist
* [x] Tests pass on CI and coverage remains at 100%
* [x] Documentation reflects the changes
* [x] No performance deterioration (if applicable)
* [x] `HISTORY.rst` has been updated
  * if this is the first change since a release, please add a new section
  * include the issue number or this pull request number `#<number>`
  * if you're not a regular contributer please include your github username `@whatever`