@gangefors 2018-09-10
No rush. For now we copy with deepcopy(). Then when you have deployed a new version we can update our code to use the new api and do a version bump.